backend/flask/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
numpy>=1.26.0
pandas>=2.2.0
scikit-learn>=1.4.0
langchain>=0.2.0
langchain-google-genai>=2.0.0
google-generativeai>=0.8.0
google-genai>=0.8.0
python-dotenv>=1.0.0
